Updates to Master
Updated the In-Game Menu within Education and Wildlife Sciences to reference specific text elements within the scene.  OpenSpot fixed for the Education level.  One bug remaining is that the Wildlife Sciences level does not appear to like the "PlaceTower" script when clicking on an OpenSpot.  Receiving an error that the CanPlaceStarbucks() function call returns a no GameObject reference when attempting to place a Starbucks on an OpenSpot.  Something that we will need to continue to work on.